requirements.txt
python-docx
google-generativeai
